Add encodable conformance to response object
