+++ AdyenNetworking/APIClient/Request.swift
@@ -63,37 +63,3 @@ public protocol Request: Encodable {
 public protocol AsyncDownloadRequest: Request {
     var onProgressUpdate: ((_ progress: Double) -> Void)? { get }
 }
-
-/// Describes an API response.
-public protocol Response: Decodable { }
-
-/// Represents an API download response.
-///
-/// The `url` property provides the temporary path to the downloaded file.
-public struct DownloadResponse: Response {
-    public let url: URL
-    
-    public init(url: URL) {
-        self.url = url
-    }
-    
-    enum CodingKeys: CodingKey {
-        case url
-    }
-    
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.url = try container.decode(URL.self, forKey: .url)
-    }
-}
-
-/// Represents an empty API response.
-public struct EmptyResponse: Response {
-    public init() { }
-}
-
-/// Describes an API Error response.
-public protocol ErrorResponse: Response, Error { }
-
-/// Represents an empty API Error response.
-public struct EmptyErrorResponse: ErrorResponse { }
